improvements on ultrasonic

+++ src/miscDevices/HcSr04KalmanFilter.cpp
@@ -5,7 +5,7 @@ HcSr04KalmanFilter::HcSr04KalmanFilter()
     /* default ctor */
 }
 
-HcSr04KalmanFilter::HcSr04KalmanFilter(double init_d, double init_v, double init_dt, double q_d, double q_v, double r)
+HcSr04KalmanFilter::HcSr04KalmanFilter(float init_d, float init_v, float init_dt, float q_d, float q_v, float r)
 {
     _dt = init_dt;
     _velocity = init_v;
@@ -44,17 +44,17 @@ HcSr04KalmanFilter::~HcSr04KalmanFilter()
 
 void HcSr04KalmanFilter::_predictionPhase()
 {
-    double d_pred = F[0][0] * _distance + F[0][1] * _velocity;
-    double v_pred = F[1][0] * _distance + F[1][1] * _velocity;
+    float d_pred = F[0][0] * _distance + F[0][1] * _velocity;
+    float v_pred = F[1][0] * _distance + F[1][1] * _velocity;
 
     _distance = d_pred;
     _velocity = v_pred;
 
     // update covariance Matrix
-    double P00 = F[0][0] * P[0][0] + F[0][1] * P[1][0];
-    double P01 = F[0][0] * P[0][1] + F[0][1] * P[1][1];
-    double P10 = F[1][0] * P[0][0] + F[1][1] * P[1][0];
-    double P11 = F[1][0] * P[0][1] + F[1][1] * P[1][1];
+    float P00 = F[0][0] * P[0][0] + F[0][1] * P[1][0];
+    float P01 = F[0][0] * P[0][1] + F[0][1] * P[1][1];
+    float P10 = F[1][0] * P[0][0] + F[1][1] * P[1][0];
+    float P11 = F[1][0] * P[0][1] + F[1][1] * P[1][1];
 
     P[0][0] = P00 + Q[0][0];
     P[0][1] = P01 + Q[0][1];
@@ -62,13 +62,13 @@ void HcSr04KalmanFilter::_predictionPhase()
     P[1][1] = P11 + Q[1][1];
 }
 
-void HcSr04KalmanFilter::_correctionPhase(double z)
+void HcSr04KalmanFilter::_correctionPhase(float z)
 {
     // calc innovation
-    double y = z - (H[0] * _distance + H[1] * _velocity);
+    float y = z - (H[0] * _distance + H[1] * _velocity);
 
     // innovation covariance
-    double S = H[0] * P[0][0] + H[1] * P[1][0] + R;
+    float S = H[0] * P[0][0] + H[1] * P[1][0] + R;
 
     // kalman-gain
     K[0] = P[0][0] / S;
@@ -79,10 +79,10 @@ void HcSr04KalmanFilter::_correctionPhase(double z)
     _velocity = _velocity + K[1] * y;
 
     // covariance correction
-    double P00 = P[0][0] - K[0] * H[0] * P[0][0];
-    double P01 = P[0][1] - K[0] * H[0] * P[0][1];
-    double P10 = P[1][0] - K[1] * H[1] * P[1][0];
-    double P11 = P[1][1] - K[1] * H[1] * P[1][1];
+    float P00 = P[0][0] - K[0] * H[0] * P[0][0];
+    float P01 = P[0][1] - K[0] * H[0] * P[0][1];
+    float P10 = P[1][0] - K[1] * H[1] * P[1][0];
+    float P11 = P[1][1] - K[1] * H[1] * P[1][1];
 
     P[0][0] = P00;
     P[0][1] = P01;
@@ -90,16 +90,16 @@ void HcSr04KalmanFilter::_correctionPhase(double z)
     P[1][1] = P11;
 }
 
-void HcSr04KalmanFilter::setVelocity(double speedMps)
+void HcSr04KalmanFilter::setVelocity(float speedMps)
 {
     this->_velocity = speedMps;
 }
-void HcSr04KalmanFilter::setDt(double dt)
+void HcSr04KalmanFilter::setDt(float dt)
 {
     this->F[0][1] = -1 * dt;
 }
 
-void HcSr04KalmanFilter::update(double z, double dt)
+void HcSr04KalmanFilter::update(float z, float dt)
 {
     setDt(dt);
     _predictionPhase();
